% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_gelu}
\alias{activation_gelu}
\title{Gaussian Error Linear Unit (GeLU) Activation Function}
\usage{
activation_gelu(x)
}
\arguments{
\item{x}{A numeric vector or matrix of pre-activation inputs.}
}
\value{
The result of the GeLU function.
}
\description{
Computes the GeLU function: $f(x) = x \cdot \Phi(x)$, where $\Phi(x)$ is the
cumulative distribution function (CDF) of the standard normal distribution.
}
