% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_relu}
\alias{activation_relu}
\title{Rectified Linear Unit (ReLU) Activation Function}
\usage{
activation_relu(x)
}
\arguments{
\item{x}{A numeric vector or matrix of pre-activation inputs.}
}
\value{
The result of the ReLU function.
}
\description{
Computes the ReLU function: $f(x) = \max(0, x)$.
}
