% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_tanh}
\alias{activation_tanh}
\title{Hyperbolic Tangent (Tanh) Activation Function}
\usage{
activation_tanh(x)
}
\arguments{
\item{x}{A numeric vector or matrix of pre-activation inputs.}
}
\value{
The result of the Tanh function, guaranteed to be a matrix.
}
\description{
Computes the Tanh function: $f(x) = \frac{e^x - e^{-x}}{e^x + e^{-x}}$.
}
