% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_tanh_grad}
\alias{activation_tanh_grad}
\title{Tanh Backward Pass (Derivative)}
\usage{
activation_tanh_grad(output, grad_output)
}
\arguments{
\item{output}{Output of the forward pass.}

\item{grad_output}{Gradient of the loss w.r.t the output.}
}
\value{
Gradient w.r.t the input, as a matrix.
}
\description{
Computes the derivative of the Tanh function: \eqn{f'(x) = 1 - f(x)^2}.
}
