% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{initialize_xavier}
\alias{initialize_xavier}
\title{Xavier Initialization}
\usage{
initialize_xavier(input_features, output_features)
}
\arguments{
\item{input_features}{The number of input connections (n_in).}

\item{output_features}{The number of output connections (n_out).}
}
\value{
A matrix of initialized weights.
}
\description{
Initializes weights using the Xavier uniform distribution.
}
