% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{initialize_zeros}
\alias{initialize_zeros}
\title{Initialize Weights to Zeros}
\usage{
initialize_zeros(input_features, output_features)
}
\arguments{
\item{input_features}{The number of input connections}

\item{output_features}{The number of output connections}
}
\value{
A matrix of zeros.
}
\description{
Initializes all weight matrix W elements to zero.
}
