% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear.R
\name{Linear}
\alias{Linear}
\title{Linear Layer Module}
\usage{
Linear(input_features, output_features)
}
\arguments{
\item{input_features}{The size of the input vector.}

\item{output_features}{The size of the output vector}
}
\value{
A list representing the Linear Module with initialized weights and biases.
}
\description{
Creates a fully connected layer module.
}
