% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequential.R
\name{Sequential}
\alias{Sequential}
\title{Sequential Module}
\usage{
Sequential(...)
}
\arguments{
\item{...}{A series of MinML modules.}
}
\value{
A list representing the Sequential Module.
}
\description{
Creates a sequential container module that stacks multiple layers
and activations together.
}
