% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimizers.R
\name{sgd_update_params}
\alias{sgd_update_params}
\title{SGD Parameter Update Step}
\usage{
sgd_update_params(current_params, gradients, learning_rate)
}
\arguments{
\item{current_params}{A named list of the current model parameters.}

\item{gradients}{A named list of the calculated gradients.}

\item{learning_rate}{The step size (alpha).}
}
\value{
A new list of updated parameters.
}
\description{
Applies the SGD update rule (theta_new = theta_old - lr * grad)
to a list of parameters using a corresponding list of gradients.
}
