% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Losses.R
\name{squared_loss}
\alias{squared_loss}
\title{Squared Loss (MSE)}
\usage{
squared_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector of the true values.}

\item{y_pred}{A numeric vector of the predicted values.}
}
\value{
The calculated MSE as a single numeric value.
}
\description{
Calculates the Mean Squared Error (MSE) between true values (y_true)
and predicted values (y_pred).
}
