% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Losses.R
\name{squared_loss_grad}
\alias{squared_loss_grad}
\title{Squared Loss (MSE) Gradient}
\usage{
squared_loss_grad(y_true, y_pred)
}
\arguments{
\item{y_true}{A numeric vector of the true values.}

\item{y_pred}{A numeric vector of the predicted values.}
}
\value{
The gradient matrix/vector.
}
\description{
Calculates the gradient of the MSE loss with respect to the prediction.
}
