% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_gelu}
\alias{activation_gelu}
\title{GeLU Activation Function}
\usage{
activation_gelu(x)
}
\arguments{
\item{x}{A numeric vector or matrix of pre-activation inputs.}
}
\value{
The result of the GeLU function as a matrix.
}
\description{
Computes GeLU: \eqn{f(x) = x * \Phi(x)}, where \eqn{\Phi(x)} is the CDF of standard normal.
}
