% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_sigmoid}
\alias{activation_sigmoid}
\title{Sigmoid Activation Function}
\usage{
activation_sigmoid(x)
}
\arguments{
\item{x}{A numeric vector or matrix of pre-activation inputs.}
}
\value{
The result of the sigmoid function.
}
\description{
Computes the Sigmoid function: $f(x) = \frac{1}{1 + e^{-x}}$.
}
