% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_sigmoid_grad}
\alias{activation_sigmoid_grad}
\title{Sigmoid Backward Pass (Derivative)}
\usage{
activation_sigmoid_grad(output, grad_output)
}
\arguments{
\item{output}{Output of the forward pass.}

\item{grad_output}{Gradient of the loss w.r.t the output.}
}
\value{
Gradient w.r.t the input, as a matrix.
}
\description{
Computes the derivative of the Sigmoid function: \eqn{f'(x) = f(x) * (1 - f(x))}.
}
