% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{activation_tanh_grad}
\alias{activation_tanh_grad}
\title{Tanh Backward Pass (Derivative)}
\usage{
activation_tanh_grad(output, grad_output)
}
\arguments{
\item{output}{The output of the forward pass, $f(x)$.}

\item{grad_output}{The gradient of the loss with respect to the output.}
}
\value{
The gradient with respect to the input, $x$.
}
\description{
Computes the derivative of the Tanh function: $f'(x) = 1 - f(x)^2$.
}
